consumer/requirements.txt
kafka-python
requests 
google-genai 